Creates a ArrowStreamDataset.

An Arrow Dataset for reading record batches from an input stream. Currently supported input streams are a socket client or stdin.

arrow_stream_dataset(host, columns, output_types, output_shapes = NULL)

Arguments

host
A tf.string tensor or string defining the input stream. For a socket client, use "<HOST_IP>:<PORT>", for stdin use "STDIN".

columns
A list of column indices to be used in the Dataset.

output_types
Tensor dtypes of the output tensors.

output_shapes
TensorShapes of the output tensors or NULL to infer partial.

Examples

if (FALSE) {
dataset <- arrow_stream_dataset(
    host,
    columns = reticulate::tuple(0L, 1L),
    output_types = reticulate::tuple(tf$int32, tf$float32),
    output_shapes = reticulate::tuple(list(), list())) %>%
  dataset_repeat(1)

sess <- tf$Session()
iterator <- make_iterator_one_shot(dataset)
next_batch <- iterator_get_next(iterator)

until_out_of_range({
  batch <- sess$run(next_batch)
  print(batch)
})
}